feat(settle): 新增订单结算功能及相关模块支持
- 新增 SettleService 及结算相关 BO、View、Converter
- 重构 CartService 商品价格计算逻辑
- 优化 SkuEntity 价格字段类型为 BigDecimal
- 重命名 CommentsService 为 CommentService
- 新增 OrderController 提供结算接口

+++ src/main/kotlin/com/hello/hello_store_service/model/view/activity/UserCouponView.kt
@@ -28,7 +28,7 @@ data class UserCouponView(
                 value = bo.getValue(),
                 tag = bo.coupon.tag,
                 desc = bo.coupon.description,
-                base = bo.coupon.threshold,
+                base = bo.coupon.threshold.toDouble(),
                 title = bo.coupon.title,
                 timeLimit = timeLimit,
             )
